Highly modified https://github.com/xfire/pydzen
This is a highly modified Pydzen. The orignal can be found
at https://github.com/xfire/pydzen.
I have created this as a wrapper for Dzen.
It is currently not working well, and needs much tweeking.

Changes from the original(off the top of my head) a so far:
    * Plugins no longer need seperate files in plugins/
    * Plugins no longer polled in the order they are found
        in config.PLUGINS. They now return data on their
        own time.
    * config.ORDER in Pydzenrc is used to determine plugin
        order.
    * Plugins now run in seperate processes.
    * Plugins now return their data to a multiprocess.Queue
    * Output to dzen is now piped through a script, as seen
        in pydzen. I could not get it working through the
        original method of using subprocess.PIPE. For now
        this method will suffice.
    * There may be more changes I cannot remember...

I made the mistake of copying my project folder without the
git repo so log history is pretty much lost.