Dynamic pathing support.
Instead of using the `HOME` environmental variable and appending
`./pydzen` to set the pydzen folder, pydzen now uses `os.getcwd()` so
that pydzen can be run in any folder and still work correctly. Of course
this means that any scripts attempting to run pydzen will likely have to
`cd` into it's folder before doing so.